# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SerialMemory API",
    "description": "Temporal knowledge graph memory system with semantic search, entity extraction, and multi-hop reasoning.",
    "contact": {
      "name": "SerialMemory",
      "url": "https://serialmemory.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/sessions/{sessionId}/end": {
      "post": {
        "tags": [
          "SerialMemory.Api"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/shadow/branches": {
      "post": {
        "tags": [
          "SerialMemory.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShadowBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "SerialMemory.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/shadow/branches/{branchId}": {
      "get": {
        "tags": [
          "SerialMemory.Api"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Shadow Branches"
        ],
        "description": "Discards a shadow branch",
        "operationId": "DiscardBranch",
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/shadow/branches/{branchId}/promote": {
      "post": {
        "tags": [
          "SerialMemory.Api"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/shadow/branches/{branchId}/freeze": {
      "post": {
        "tags": [
          "SerialMemory.Api"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/shadow/branches/create": {
      "post": {
        "tags": [
          "Shadow Branches"
        ],
        "description": "Creates a new shadow branch",
        "operationId": "CreateBranch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateBranchApiRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "branchType": {
            "type": "string",
            "nullable": true
          },
          "parentBranchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sourceMemoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "confidence": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "hypothesis": {
            "type": "string",
            "nullable": true
          },
          "reasoningContext": {
            "type": "string",
            "nullable": true
          },
          "ttlHours": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "autoDiscardThreshold": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ShadowBranchRequest": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "branchType": {
            "type": "string",
            "nullable": true
          },
          "hypothesis": {
            "type": "string",
            "nullable": true
          },
          "reasoningContext": {
            "type": "string",
            "nullable": true
          },
          "sourceMemoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "confidence": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "ttlHours": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "autoDiscardThreshold": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API key for authentication",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}